[Shared] Extract models to shared package
