+++ packages/app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "UICompanion",
+  "name": "uicompanion-app",
   "version": "1.0.0",
   "scripts": {
     "build": "tsc --build",
@@ -7,6 +7,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "@ls1intum/uicompanion-shared": "^1.0.0",
     "axios": "^1.4.0",
     "cors": "^2.8.5",
     "express": "^4.18.2",
