--- packages/app/src/models/IssueStatus.ts
@@ -1,6 +1,0 @@
-export enum IssueStatus {
-    OPEN = 'Open',
-    IN_PROGRESS = 'In Progress',
-    IN_REVIEW = 'In Review',
-    APPROVED = 'Approved'
-}
