--- packages/plugin/src/app/models/Issue.ts
@@ -1,9 +1,0 @@
-export interface Issue {
-    repository_url: string;
-    number: number;
-    title: string;
-    description: string;
-    status: string;
-    frames: string[];
-    prototypeUrls: string[];
-}
