+++ packages/server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "uicompanionserver",
+  "name": "uicompanion-server",
   "version": "1.0.0",
   "description": "",
   "scripts": {
@@ -13,6 +13,7 @@
     "typescript": "^5.1.6"
   },
   "dependencies": {
+    "@ls1intum/uicompanion-shared": "^1.0.0",
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
