+++ packages/server/src/model/Issue.ts
@@ -1,46 +1,4 @@
-import { Document, Schema, Model, model, Error } from 'mongoose'
+import { Model, model } from "mongoose";
+import { IIssue, issueSchema } from "@ls1intum/uicompanion-shared/models/Issue";
 
-export interface IIssue extends Document {
-    repository_url: string;
-    number: number;
-    title: string;
-    description: string;
-    status: string;
-    frames: string[];
-    prototypeUrls: string[];
-}
-
-const issueSchema = new Schema({
-    repository_url: {
-        type: String,
-        required: true,
-    },
-    number: {
-        type: Number,
-        required: true,
-        unique: true,
-    },
-    title: {
-        type: String,
-        required: true,
-    },
-    description: {
-        type: String,
-        required: false,
-    },
-    status: {
-        type: String,
-        required: true,
-        enum: ['Open', 'In Progress', 'In Review', 'Approved'],
-    },
-    frames: {
-        type: [String],
-        required: true,
-    },
-    prototypeUrls: {
-        type: [String],
-        required: true,
-    },
-});
-
-export const Issue: Model<IIssue> = model<IIssue>('Issue', issueSchema);
+export const Issue: Model<IIssue> = model<IIssue>('Issue', issueSchema);
