+++ packages/uicompanion-shared/README.md
@@ -1,0 +1,15 @@
+# UICompanion Shared
+
+This is a comprehensive npm package serving as the core foundation for the UICompanion services. This package is designed to consolidate common functionalities, models, and utilities that are utilized across the other three UICompanion services. By centralizing shared resources, UICompanion Shared ensures consistency, reduces redundancy, and facilitates easier maintenance and updates across the UICompanion ecosystem.
+
+## Installation
+```bash
+npm install @ls1intum/uicompanion-shared
+```
+
+## Usage
+Import the required models, functions, or utilities in your UICompanion service project:
+
+```Typescript
+const { Model, UtilityFunction } = require('uicompanion-shared');
+```
