+++ packages/uicompanion-shared/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@ls1intum/uicompanion-shared",
+  "version": "1.0.0",
+  "description": "",
+  "repository": "https://github.com/ls1intum/UICompanion.git",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "mongoose": "^8.0.2"
+  },
+  "publishConfig": {
+    "registry": "https://npm.pkg.github.com/ls1intum"
+  }
+}
