+++ packages/uicompanion-shared/tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+        "declaration": true,
+        "declarationMap": true,
+        "emitDeclarationOnly": true,
+        "esModuleInterop": true,
+        "lib": [
+            "dom",
+            "es2015",
+            "es2016",
+            "es2017",
+            "es2018",
+            "es2019",
+            "es2020"
+        ],
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "target": "es2020",
+        // Enhance strictness.
+        "noImplicitReturns": true,
+        "noUnusedLocals": true,
+        "strict": true,
+    }
+}
